Implement connection handling
This adds the majority of the logic for subscription connection
handling. It delegates the actual authentication and resolving of the
subscription to the user application but it properly implements the
GraphQL WS protocol as it's expected by clients.

This is still a work in progress because authentication currently needs
to happen synchronously. However, a database call may be desired to
perform authentication, so it should be possible to do this
asynchronously. That would be a breaking change to the current API.